Add support for image upload to Medium
In order for local images to be used in Medium articles, they must first
be uploaded to Medium. This is accomplished using Medium's image upload
api endpoint. The local path is replaced with the image's url on Medium
in the markdown source so it is referenced properly online. Remote
images are left alone since Medium handles those automatically.